\name{c2}
\docType{data}
\alias{c2}
\title{Default vector of classes2}
\description{
   This is a default classes vector, used to show the usage of "getErrors". This is a dummy object, created randomly with some random classes. A simple vector of integers with values  between 0 and 2.
}
\usage{data(c2)}
\format{A vector of classes.}
\source{Alvaro Gonzalez Alvarez}
\keyword{datasets}

