\name{getClasses}
\alias{gC}
\alias{getClasses}
\title{Returns the classes of a given data}
\description{Returns the classes of a given data. The class vector must be the last column.}
\usage{
 getClasses(data)
}
\arguments{

  Takes the data.

  \item{data}{The dataset which contain the classes.}
  
}

\value{The classes of the problem.}

\examples{
 data(trainA)
 getClasses(trainA)
}


\keyword{univar}
