\name{Glass}
\docType{data}
\alias{glass}
\title{Glass Identification Database}
\description{
  This is a matrix with 10 columns. Each row is a point, there are 214 points. The 9 first columns are the value of a dimension (or attribute) of the point, the 10th is the class of that point. The attributes are:

\begin{enumerate}
\item RI: refractive index
   \item Na: Sodium (unit measurement: weight percent in corresponding oxide, as 
                  are attributes 4-10)
   \item Mg: Magnesium
   \item Al: Aluminum
   \item Si: Silicon
   \item K: Potassium
   \item Ca: Calcium
   \item Ba: Barium
   \item Fe: Iron
\end{enumerate}
}
\usage{data(glass)}
\format{A matrix.}
\source{
\begin{itemize}
\item UCI ML Repository
\item Creator: B. German
 \item Central Research Establishment
           Home Office Forensic Science Service
           Aldermaston, Reading, Berkshire RG7 4PN
\item Donor: Vina Spiehler, Ph.D., DABFT
               Diagnostic Products Corporation
               (213) 776-0180 (ext 3014)
 \item Date: September, 1987
\end{itemize}
}
\keyword{datasets}

