\encoding{latin1}
\name{Iris-test}
\docType{data}
\alias{iris.test}
\title{Test problem for IRIS}
\description{
  This is a matrix with 5 columns. Each row is a point, there are 15 points. The 5 first columns are the value of a dimension of the point, the 5th is the class of that point.  It is a premade testing set.
}
\usage{data(iris.test)}
\format{A matrix.}
\source{Teaching Materials, Laboratory Classes of
Knowledge Engineering, Luciano Sánchez, EPSI Gijón}
\keyword{datasets}
