\encoding{latin1}
\name{Iris-train}
\docType{data}
\alias{iris.train}
\title{Train problem for IRIS}
\description{
  This is a matrix with 5 columns. Each row is a point, there are 135. The 4 first columns are the value of a dimension of the point, the 5th is the class of that point. It is a premade training set.
}
\usage{data(iris.train)}
\format{A matrix.}
\source{Teaching Materials, Laboratory Classes of
Knowledge Engineering, Luciano Sánchez, EPSI Gijón}
\keyword{datasets}

