\name{Iris}
\docType{data}
\alias{IRIS}
\title{Iris Plant Database}
\description{
  This is a matrix with 5 columns. Each row is a point, there are 150 points. The 4 first columns are the value of a dimension (or attributes) of the point, the 5th is the class of that point. The attributes are:

\begin{enumerate}
\item sepal length in cm
 \item sepal width in cm
\item petal length in cm
 \item petal width in cm
	\end{enumerate}
}
\usage{data(IRIS)}
\format{A matrix.}
\source{

\begin{itemize}
 	\item UCI ML Repository
 \item Creator: R.A. Fisher
 \item Donor: Michael Marshall (MARSHALL%PLU@io.arc.nasa.gov)
 \item Date: July, 1988
\end{itemize}
}
\keyword{datasets}

