\name{mt}
\docType{data}
\alias{mt}
\title{Default Meta Information}
\description{
  This is a default meta information list. It is used to show the usage 
of 'KbtoXML'. The list has this values:

The Author:
\begin{ExampleCode}
$Author = "John Doe" 
\end{ExampleCode}

The Description:
\begin{ExampleCode}
$Description= "Test fS"
\end{ExampleCode}

}
\usage{data(mt)}
\format{A meta information list.}
\source{Alvaro Gonzalez Alvarez}
\keyword{datasets}

