\name{Pima}
\docType{data}
\alias{pima}
\title{Pima Indians Diabetes Database}
\description{
  This is a matrix with 9 columns. Each row is a point, there are 768 points. The 8 first columns are the value of a dimension (or attributes) of the point, the 9th is the class of that point. The attributes are:

\begin{enumerate}
 \item Number of times pregnant
   \item Plasma glucose concentration a 2 hours in an oral glucose tolerance test
   \item Diastolic blood pressure (mm Hg)
   \item Triceps skin fold thickness (mm)
   \item 2-Hour serum insulin (mu U/ml)
   \item Body mass index (weight in kg/(height in m)\eqn{^2}{^2})
   \item Diabetes pedigree function
   \item Age (years)
\end{enumerate}

}

\source{
 \begin{itemize}
\item UCI ML Repository
\item Original owners: National Institute of Diabetes and Digestive and
                        Kidney Diseases
\item Donor of database: Vincent Sigillito (vgs@aplcen.apl.jhu.edu)
                          Research Center, RMI Group Leader
                          Applied Physics Laboratory
                          The Johns Hopkins University
                          Johns Hopkins Road
                          Laurel, MD 20707
                          (301) 953-6231
\item Date received: 9 May 1990
 \end{itemize}
}
\usage{data(pima)}
\format{A matrix.}
\keyword{datasets}

