\encoding{latin1}
\name{Pl}
\docType{data}
\alias{Pl}
\title{Vector of partitions for testing IRIS}
\description{
  This is a list of partitions, suitable for problem IRIS.
}
\usage{data(Pl)}
\format{A vector of partitions.}
\source{Teaching Materials, Laboratory Classes of
Knowledge Engineering, Luciano Sánchez, EPSI Gijón}
\keyword{datasets}

