\encoding{latin1}
\name{TrainA}
\docType{data}
\alias{trainA}
\title{Problem A}
\description{
  This is a matrix with 4 columns. Each row is a point, there are 200 points. The 3 first columns are the value of a dimension of the point, the 4th is the class of that point. This is a dummy example used as an example at class lectures. The A/B/C/D/M problems has a deliberate variability to show how different problems act when faced to different algorithms.
}
\usage{data(trainA)}
\format{A matrix.}
\source{Teaching Materials, Laboratory Classes of
Knowledge Engineering, Luciano Sánchez, EPSI Gijón}
\keyword{datasets}
\encoding{latin1}
