\encoding{latin1}
\name{TrainB}
\docType{data}
\alias{trainB}
\title{Problem B}
\description{
  This is a matrix with 3 columns. Each row is a point, there are 4001 points. The 2 first columns are the value of a dimension of the point, the 3rd is the class of that point. This is a dummy example used as an example at class lectures. The A/B/C/D/M problems has a deliberate variability to show how different problems act when faced to different algorithms.

}
\usage{data(trainB)}
\format{A matrix.}
\source{Teaching Materials, Laboratory Classes of
Knowledge Engineering, Luciano Sánchez, EPSI Gijón}
\keyword{datasets}

