\encoding{latin1}
\name{TrainC}
\docType{data}
\alias{trainC}
\title{Problem C}
\description{
  This is a matrix with 8 columns. Each row is a point, there are 768 points. The 7 first columns are the value of a dimension of the point, the 8th is the class of that point. This is a dummy example used as an example at class lectures. The A/B/C/D/M problems has a deliberate variability to show how different problems act when faced to different algorithms.

}
\usage{data(trainC)}
\format{A matrix.}
\source{Teaching Materials, Laboratory Classes of
Knowledge Engineering, Luciano Sánchez, EPSI Gijón}
\keyword{datasets}

