\encoding{latin1}
\name{TrainD}
\docType{data}
\alias{trainD}
\title{Problem D}
\description{
  This is a matrix with 5 columns. Each row is a point, there are 200 points. The 4 first columns are the value of a dimension of the point, the 5th is the class of that point. This is a dummy example used as an example at class lectures. The A/B/C/D/M problems has a deliberate variability to show how different problems act when faced to different algorithms.
}
\usage{data(trainD)}
\format{A matrix.}
\source{Teaching Materials, Laboratory Classes of
Knowledge Engineering, Luciano Sánchez, EPSI Gijón}
\keyword{datasets}

