\encoding{latin1}
\name{TrainM}
\docType{data}
\alias{trainM}
\title{Problem Minimum}
\description{
  This is a matrix with 5 columns. Each row is a point, there are 10 points. The 4 first columns are the value of a dimension of the point, the 5th is the class of that point. It is a minimum problem, created for exampling and debugging purposes. This is a dummy example used as an example at class lectures. The A/B/C/D/M problems has a deliberate variability to show how different problems act when faced to different algorithms.
}
\usage{data(trainM)}
\format{A matrix.}
\source{Teaching Materials, Laboratory Classes of
Knowledge Engineering, Luciano Sánchez, EPSI Gijón}
\keyword{datasets}

