\name{Wine}
\docType{data}
\alias{wine}
\title{Wine Recognition Database}
\description{
  This is a matrix with 14 columns. Each row is a point, there are 176 points. The 13 first columns are the value of a dimension (or attribute) of the point, the 14th is the class of that point. The attributes are:

 \begin{enumerate}
	 \item Alcohol
	 \item Malic acid
 	 \item Ash
	 \item Alcalinity of ash  
 	 \item Magnesium
	 \item Total phenols
 	 \item Flavanoids
 	 \item Nonflavanoid phenols
 	 \item Proanthocyanins
	 \item Color intensity
 	 \item Hue
 	 \item OD280/OD315 of diluted wines
 	 \item Proline       
 \end{enumerate}
}
\usage{data(wine)}
\format{A matrix.}


\source{
\begin{itemize}
\item UCI ML Repository
\item  Forina, M. et al, PARVUS - An Extendible Package for Data
       Exploration, Classification and Correlation. Institute of Pharmaceutical
       and Food Analysis and Technologies, Via Brigata Salerno, 
       16147 Genoa, Italy.

\item Stefan Aeberhard, email: stefan@coral.cs.jcu.edu.au
\item July 1991

 \end{itemize}
}
\keyword{datasets}

