\name{XML2kb}
\alias{XML2kb}
\title{Reads a knowledge base from a XML}
\description{Reads a knowledge base from a XML file.}
\usage{
 XML2kb(file)
}
\arguments{

  Takes the name of the XML file.

  \item{file}{The name of the XML file.}
  
}

\value{Returns the knowledge base, if success. FALSE 0, is returned otherwise.}

\examples{
 
 #XML2kb("sb.xml")
}


\keyword{univar}
